Disable the syslog long message over Unix datagram socket test on macOS
